FINISH include typescript 5/5 - Blog page
